Package and variables refactor
